+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K0RTHHQ1Y2T7Z7P69GB1A4YX"
+        "01K0V7JH20BB3CSPZTCJXSFXYH"
     )
     r"""The id of the resource"""