+++ packages/orq-rc/src/orq_ai_sdk/models/deployments.py
@@ -266,14 +266,14 @@ r"""The type of the content part. Always `file`."""
 
 class FileTypedDict(TypedDict):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
 
 
 class File(BaseModel):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
 
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
@@ -723,14 +723,14 @@ r"""The type of the content part. Always `file`."""
 
 class TwoFileTypedDict(TypedDict):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
 
 
 class TwoFile(BaseModel):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
 
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""