+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K0RTHHQ0V227ASD5PPRFCGKY"
+        "01K0V7JH1ZD6D9TXZRVKBQCTAW"
     )
     r"""The id of the resource"""