+++ packages/orq-rc/src/orq_ai_sdk/models/retrievecontactop.py
@@ -101,7 +101,7 @@ class RetrieveContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-07-22T10:23:13.475Z")
+    updated: Optional[datetime] = parse_datetime("2025-07-23T08:49:25.849Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")