+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -288,14 +288,14 @@ r"""The type of the content part. Always `file`."""
 
 class RetrieveDatapoint2FileTypedDict(TypedDict):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
 
 
 class RetrieveDatapoint2File(BaseModel):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
 
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
@@ -572,5 +572,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-07-22T10:23:13.475Z")
+    updated: Optional[datetime] = parse_datetime("2025-07-23T08:49:25.849Z")
     r"""The date and time the resource was last updated"""