+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatapointop.py
@@ -276,14 +276,14 @@ r"""The type of the content part. Always `file`."""
 
 class UpdateDatapoint2FileTypedDict(TypedDict):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
 
 
 class UpdateDatapoint2File(BaseModel):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
 
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
@@ -813,14 +813,14 @@ r"""The type of the content part. Always `file`."""
 
 class UpdateDatapoint2DatasetsFileTypedDict(TypedDict):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
 
 
 class UpdateDatapoint2DatasetsFile(BaseModel):
     file_data: str
-    r"""The base64 encoded file data, used when passing the file to the model as a string."""
+    r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
 
     filename: str
     r"""The name of the file, used when passing the file to the model as a string."""
@@ -1098,5 +1098,5 @@ class UpdateDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-07-22T10:23:13.475Z")
+    updated: Optional[datetime] = parse_datetime("2025-07-23T08:49:25.849Z")
     r"""The date and time the resource was last updated"""