+++ packages/orq-rc/RELEASES.md
@@ -2059,3 +2059,13 @@ Based on:
 - [python v3.10.0-rc.14] packages/orq-rc
 ### Releases
 - [PyPI v3.10.0-rc.14] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.14 - packages/orq-rc
+
+## 2025-07-23 08:51:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.586.0 (2.661.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.10.0-rc.16] packages/orq-rc
+### Releases
+- [PyPI v3.10.0-rc.16] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.16 - packages/orq-rc